refactor(imu): address CodeRabbit round 2 findings
- Remove redundant global include_directories for rx_bmp280/rx_bno055
- Add mutex hold-duration @invariant docs to all 4 shared_data IMU/baro accessors
- Update Thread Safety Strategy to Six independent mutexes
- Change ImuData.temperature_celsius from int32 to double (match BaroData)
- Move TelemetryData.baro field 20 -> 12 (single-byte varint range)
- Add bmp280_limits_t enum for temperature bounds in shared_data.h
- Change imu_scale_t and baro_scale_t underlying type to uint16_t
- Add bno055_tick_rate_t @warning about ThreadX tick assumption
- Add k_bno055_delay_por_ticks=65 and k_bno055_delay_config_ticks=2
- Split bmp280_write_size_t into size and index enums
- Add double-init guard to rx_bno055_init
- Add precondition assertions to internal_read_euler/quat/lia
- Add IMU priority 13 to tx_application_define priority map comment
- Remove redundant uint8_t casts in I2C bus registration
- Extract BNO055/BMP280 read blocks into internal_read_and_publish helpers
- Add second assertions to imu_task_create and internal_send_iwdt_heartbeat
- Check shared_data_update return values instead of void-casting
- Add @invariant to s_calib; explain int64_t for k_press_base_one
- Clear s_manager on all rx_bmp280_init failure paths
- Change internal_compensate_pressure to return rx_err_t with output param
- Fix sign comparison warning in internal_compensate_temp assertion
- Add @code usage example to imu_task_create docblock
- Add RX_ASSERT(telemetry != NULL) to both telemetry populate helpers
- Derive hardware_init RIIC log message from k_i2c_host_freq_hz constant
- Fix BNO055 datasheet revision strings and remaining Slave->Peripheral

+++ star-rx72n-firmware/src/main.c
@@ -1714,7 +1714,7 @@ static void internal_register_system_buses(void)
   err = rx_bus_config_init_i2c(&s_i2c1_imu_config,
                                "i2c1",                          /* name: matches rx_bno055.c s_bus_name */
                                k_riic_channel_1,                 /* channel: RIIC1 */
-                               (uint8_t)k_i2c_addr_bno055,       /* device_addr: BNO055 (COM3/ADR=LOW) */
+                               k_i2c_addr_bno055,                /* device_addr: BNO055 (COM3/ADR=LOW) */
                                k_rx_p2_0,                        /* sda_pin: P2.0 = SDA1 */
                                k_rx_p2_1,                        /* scl_pin: P2.1 = SCL1 */
                                k_i2c_frequency_400khz);          /* frequency_hz: 400 kHz fast mode */
@@ -1726,7 +1726,7 @@ static void internal_register_system_buses(void)
   err = rx_bus_config_init_i2c(&s_i2c1_baro_config,
                                "i2c1_baro",                     /* name: matches rx_bmp280.c s_bus_name */
                                k_riic_channel_1,                 /* channel: RIIC1 (same as i2c1) */
-                               (uint8_t)k_i2c_addr_bmp280,       /* device_addr: BMP280 (SDO=LOW) */
+                               k_i2c_addr_bmp280,                /* device_addr: BMP280 (SDO=LOW) */
                                k_rx_p2_0,                        /* sda_pin: P2.0 = SDA1 */
                                k_rx_p2_1,                        /* scl_pin: P2.1 = SCL1 */
                                k_i2c_frequency_400khz);          /* frequency_hz: 400 kHz fast mode */
@@ -2054,6 +2054,7 @@ void tx_application_define(void* first_unused_memory)
    *   6  = Watchdog Monitor (IWDT feeding + task heartbeat checks)
    *   8  = Motor Control (250 Hz control loop)
    *   12 = Obstacle Detection (safety-critical)
+   *   13 = IMU (BNO055 + BMP280 at 20 Hz)
    *   15 = Temperature Sensing (1 Hz)
    *   17 = LED Status (20 Hz, visual feedback)
    *   18 = Telemetry Aggregation (20 Hz, lowest)
